apps/pipeline/requirements.txt
edgartools[ai]>=5.28
finnhub-python>=2.4
yfinance>=0.2
supabase>=2.0
python-dotenv>=1.0
pydantic>=2.0
requests>=2.31
tweepy>=4.14
